Restructure symbols in ~/internals/data-source-helper
Create new files for separate helper functions
