+++ packages/server/src/internals/data-source-helper/common.ts
@@ -1,0 +1,9 @@
+/**
+ * Error which is thrown when an invalid value was returned by a data source
+ * request.
+ */
+export class InvalidValueError extends Error {
+  constructor(public invalidValue: unknown) {
+    super('Invalid value received from a data source.')
+  }
+}
